+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `SingleFileDisasm`:
-  - Error out if the end address is larger than the start address of `.text` or
+  - Error out if the start address is larger than the end address of `.text` or
     data.
   - Avoid writing a `.text.s` file if the `.text` start and end addresses are
     the same, as long as the user has given data addresses, otherwise panic.